+++ app/src/main/java/com/example/appcentnasaapi/domain/model/roverResponse/Photo.kt
@@ -1,4 +1,4 @@
-package com.example.appcentnasaapi.model.roverResponse
+package com.example.appcentnasaapi.domain.model.roverResponse
 
 
 import com.google.gson.annotations.SerializedName
